LayoutTests/http/tests/websocket/tests/hybi/multiple-subprotocols-expected.txt

Test WebSocket subprotocol negotiation.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Connected
Received: superchat
PASS receivedMessage is "superchat"
Closed
PASS closeEvent.wasClean is true
PASS successfullyParsed is true

TEST COMPLETE
